requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.14
charset-normalizer==2.1.1
Django==4.1.1
django-cors-headers==3.13.0
django-dotenv==1.4.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
docopt==0.6.2
idna==3.4
PyJWT==2.5.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
yarg==0.1.9
